# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Manager API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get all tasks",
                "description": "Returns all tasks from the system that the user has access to.",
                "operationId": "indexTasks",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Filter tasks by status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "completed",
                                "incompleted",
                                "overdue",
                                "assigned_to_me",
                                "created_by_me",
                                "assigned_to_others",
                                "created_by_others"
                            ]
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort tasks by field",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "dueDate",
                                "-dueDate",
                                "title",
                                "-title",
                                "completed",
                                "-completed",
                                "createdAt",
                                "-createdAt"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tasks retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TaskResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Create a task",
                "description": "Create a task",
                "operationId": "storeTasks",
                "requestBody": {
                    "description": "Task object that needs to be created",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Task title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Task description"
                                    },
                                    "assigned_to": {
                                        "description": "User ID or email",
                                        "type": "string",
                                        "example": "9a136676-8ebb-4304-9de9-78bc74ab1d69"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2021-01-01"
                                    },
                                    "file": {
                                        "type": "string",
                                        "format": "binary",
                                        "example": "file.pdf"
                                    },
                                    "completed": {
                                        "type": "boolean",
                                        "example": "false"
                                    },
                                    "created_by": {
                                        "type": "string",
                                        "example": "9a136676-8ebb-4304-9de9-78bc74ab1d69"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/TaskResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/tasks/{task}/file": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Download the file attached to the given task",
                "description": "Download the file attached to the given task",
                "operationId": "downloadFileTasks",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "ID of the task to download the file from",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File downloaded successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tasks/{task}/complete": {
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Mark a task as completed",
                "description": "Mark a task as completed and notify the assignee if they are not the creator.",
                "operationId": "completeTask",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "ID of the task to complete",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get all users",
                "description": "Returns a list of all users",
                "operationId": "getUsers",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "TaskResource": {
                "title": "TaskResource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "title": {
                        "type": "string",
                        "example": "Task title"
                    },
                    "description": {
                        "type": "string",
                        "example": "Task description"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-01-01T00:00:00.000000Z"
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2021-01-01"
                    },
                    "completed": {
                        "type": "boolean",
                        "example": "false"
                    },
                    "created_by": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "assigned_to": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "file": {
                        "type": "string",
                        "format": "binary",
                        "example": "http://localhost:8000/api/tasks/1/file"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "title": "UserResource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "john.doe@email.com"
                    },
                    "is_admin": {
                        "type": "boolean",
                        "example": "false"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}